fix: resolve all clippy warnings in projects and cli packages
Fixed 33 errors in orkee-projects and 1 error in orkee-cli:
- Removed unused imports (Serialize, ContextSnapshot, ManagerError, etc.)
- Fixed unused variables by prefixing with underscore
- Removed unnecessary mut qualifiers
- Resolved ambiguous glob re-exports in context/mod.rs
- Applied clippy suggestions (or_default, push char, collapsed if, etc.)
- Added allow(dead_code) for intentionally unused code

All packages now pass cargo clippy -- -D warnings with zero errors.

+++ packages/projects/src/context/history_service.rs
@@ -1,6 +1,6 @@
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
-use sqlx::{Row, SqlitePool};
+use sqlx::SqlitePool;
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct ContextSnapshot {
@@ -262,7 +262,7 @@ impl HistoryService {
             .into_iter()
             .map(|r| TokenUsage {
                 date: r.date.unwrap_or_default(),
-                tokens: r.tokens as i32,
+                tokens: r.tokens,
             })
             .collect())
     }
